chore(proto)!: fully split sequencerapis and remove (#958)
## Summary
Finishes the split up of the former `sequencer` apis into separate
modules & repositories. `primitives` becomes the core that any other
module can rely on, while `protocolapis` is shapes specifically needed
for submitting transactions or interacting with the state of the astria
blockchain. `sequencerblock` is information needed only for execution of
already sequenced data.

## Background
The ultimate goal is to decouple the write path and read path of
executing, such that future changes to transactions shapes are not
necessarily breaking changes for dependent rollups.

## Changes
- Removed sequencer protos & astria-core::sequencer::v1
- Added types to primitive and created protocol packages.

## Testing
CI/CD

## Breaking Changelist
- This is a breaking change for everything dependent on the sequencer
protos package.