chore(sequencer): update justfile and testnet script (#985)
## Summary
Updated the sequencer's justfile and testnet.sh script to work again
with latest version of sequencer.

TODO: have someone test on their linux machine. I had to edit the script
to work on a mac and tried to make the commands work on both but I'm
unsure if I did it right.

## Testing
`just run-testnet` : to run testnet locally 
`just clean-testnet` : to cleanup testnet