fix(composer): add capacity to bundle factory (#937)
## Summary
This adds logic for setting the `BundleFactory`'s `finished_queue`
capacity and for the `Executor` to stop pulling transactions from the
`serialized_rollup_transactions` channel while its `BundleFactory` is
full

## Background
The `Executor`'s `BundleFactory` has a queue of `SizedBundle`s that it
builds as transactions are collected. This was previously uncapped,
potentially leading to OOM problems (e.g. k8s shutting down the
process).

## Changes
- Added config variable for `finished_queue_capacity`
- Added variable to constructors for `Executor` and `BundleFactory`
- `BundleFactory::try_push` can now return
`BundleFactoryError:FinishedQueueFull`
- The `BundleFactory` now tracks whether it is full or not.
- `Executor` will only pull from the channel if it has available
capacity in its `BundleFactory`.

## Testing
- sanity check to see that the `BundleFactory` is indeed marked full
when it should be
- test that `BundleFactory::next_finished().pop()` on a full bundle sets
it to no longer full
- test that `BundleFactory::pop_now` on a full bundle sets it to no
longer full

## Breaking Changelist
- `Executor` will only pull from the channel if it has available
capacity in its `BundleFactory`. This will cause the collectors'
`try_send` to time out when back pressure accumulates in the system,
dropping the collected transactions.

## Related Issues
Link any issues that are related, prefer full github links.

closes <!-- list any issues closed here -->