feat(sequencer-relayer)!: submit blobs directly to celestia app (#963)
## Summary
The sequencer-relayer has been updated to submit blobs directly to the
Celestia app via its gRPC interface rather than using the Celestia node
as a proxy.

## Background
There have been issues experienced in the past when using the Celestia
node for blob submissions related to sequences and timeouts. We want to
avoid such issues by "cutting out the middle man" and controlling the
flow of submissions directly inside the relayer.

## Changes
This PR adds a lot of new proto files and their corresponding generated
Rust sources, although they have been reduced to the minimal set of
messages/rpcs required for this new feature.

Aside from these, most of the new code has been isolated to a new module
`celestia_client` inside the `relayer` module. It should be reasonably
self-contained, with a view to possibly extracting it to a separate
crate in the future, and/or providing a generic client API should we
wish to add handling for more chains in the future.

At a high level, the new client is constructed via a builder, which may
need multiple attempts to create the client in the case that the
Celestia app is unreachable. Retries are controlled via `tryhard`,
maxing out at 30s and with infinite attempts. A shutdown signal can
cancel this, allowing for graceful exit. The builder retrieves initial
info from the Celestia app (e.g. chain ID, gas costs). The immutable
chain ID is held as a member of the client, but the other values (being
variable) are cached in the `State` struct and can be retrieved and
cached again should an attempt to submit blobs fail.

The `CelestiaClient` only has one public method: `try_submit`. This is
called in the existing `relayer::write::submit_with_retry` function in
much the same way as previously. The main difference here is that the
client's `try_submit` method is also passed a tokio watch receiver
containing the optional last error. When first submitting a given batch
of blobs, the watch channel holds `None`. On each failed attempt, the
`on_retry` closure sends the error returned by `try_submit`. On a fresh
batch of blobs, a new watch channel is constructed, again initialized
with `None`. The purpose of providing the last error to the client is
that it can be interrogated for a failure due to the fee being too low.
Such an error holds information on the required fee, and on reattempt,
the client uses this value in preference to its own calculated fee.

The general flow inside `try_submit` is as follows:
- if the previous submit attempt failed, fetch the cost info from the
Celestia app and cache it
- fetch our account number and sequence number from the Celestia app
- create, sign and send the blob transaction via the Celestia
`BroadcastTx` method
- poll the Celestia app once per second via the `GetTx` method

## Testing
Most of the new functionality has unit tests. The `CelestiaClient`
mainly has member methods calling gRPC requests, however handling the
responses is delegated to free functions. As well as code clarity, this
allows for comprehensive unit tests of these functions without the need
for running/mocking a Celestia server. However, there are still missing
tests for the `try_submit` method. That might be best done via the new
`astria-grpc-mock` crate, or at a higher level by refactoring the
existing black box tests (almost all of which have been removed for
now).

I have run the sequencer against a local network running on k8s and
observed the blob submission to behave as expected via the logging. This
includes temporarily changing the fee to a low value for the first
submission attempt, which caused the retry behaviour to identify and use
the correct fee from the error response of the first attempt. I have
also had the smoke test pass locally.

## Breaking Changelist
- `ASTRIA_SEQUENCER_RELAYER_CELESTIA_BEARER_TOKEN` has been replaced by
`ASTRIA_SEQUENCER_RELAYER_CELESTIA_APP_KEY_FILE` which expects a path to
a hex-encoded secp256k1 secret key corresponding to an account on the
Celestia chain.
- `ASTRIA_SEQUENCER_RELAYER_CELESTIA_ENDPOINT` has been replaced by
`ASTRIA_SEQUENCER_RELAYER_CELESTIA_APP_GRPC_ENDPOINT` which represents
the endpoint of the gRPC server on the Celestia app.

## Further Discussion
- integration testing/black box testing. [Discussed, agreed to restore
them in follow-up. Ticketed at #1008.]
- the polling period for `GetTx` after a successful `BroadcastTx` is
hard-coded to one second. Seems like this should be configurable via an
env var. [Discussed, and the period is now variable depending upon the
result of `GetTx`. Also, given that the Celestia app will likely be
managed by us, we can control its rate-limits, meaning there's less
likelihood of needing to change the polling period in the relayer.]

## Related Issues
Closes #921.