fix(sequencer): stateful check now ensures balance for total tx (#1009)
## Summary
Sequencer's stateful check wasn't ensuring that a transaction's entire
fee consumption was covered by the balance, it was only checking per
individual action. Now the stateful check will err if there aren't
enough funds for the whole transaction.

## Background
We were checking this already in the mempool, but this check will cover
the edge case of an account's earlier transaction consuming funds in the
same block.

## Testing
Added unit test.

## Related Issues
closes #786