fix(chart): secretObject when no env var usage (#1023)
## Summary
Removes the unused `secretObjects` in relayer chart, and the `configmap`
which is unused when `secretProvider` is enabled

## Background
Secret objects are used for loading into k8s secrets and env vars, this
is not something we are doing for the relayer chart as the secret is
only accessed directly as a file.

## Changes
- disable configmap when secret provider enabled
- remove `secretObjects` from the provider.

## Testing
CI/CD, and changes have been manually made to deployments for devnet.