fix(sequencer): panic sequencer instead of cometbft on erroring abci consensus requests (#1016)
## Summary
Zellic suggested that we panic the sequencer on consensus failures
instead of silently sending them to cometbft. There isn't any error
handling in cometbft so doing this makes sense and gives us more info
for debugging.

## Changes
Made the request handler panic in the sequencer on abci requests that
shouldn't fail.

## Testing
Ran the sequencer locally with planned failures. 

## Related Issues

closes #1004