test(sequencer-relayer): reinstate black box tests (#1033)
## Summary
The black box tests were removed in a recent PR. This PR reinstates them
using the new mock gRPC framework `astria-grpc-mock`.

## Background
The tests would have needed heavy refactoring in #963, and we wanted to
avoid adding a lot more code to that already-large PR. We decided to
temporarily delete them and reinstate them using `astria-grpc-mock` to
mock responses from the Celestia app.

## Changes
The tests removed in #963 and a single test briefly added in #1001 then
removed again have been restored. I also added a new test to check the
relayer shuts down in a timely manner.

Previously the tests leaned heavily on counts of blobs received by the
mock Celestia app. The current tests retain these checks, but also query
the `/status` endpoint of the sequencer-relayer to confirm its state.
There was also a single test previously which checked the state of the
postsubmit.json file. I didn't think this was necessary given that we're
querying the state in a more "black box" way now (via the http server),
but I didn't remove the function to perform this check
(`TestSequencerRelayer::assert_state_files_are_as_expected`) pending a
decision on whether to reinstate that check or not inside the
`later_height_in_state_leads_to_expected_relay` test.

As @SuperFluffy predicted, all the new protobuf packages added in #963
had to be added to the `pbjson_build` builder to generate the serde
impls required for using them in `astria-grpc-mock`. This accounts for
the vast majority of the new LoC in this PR.

I made a few small changes to the mock framework:
- added `Mock::up_to_n_times` to avoid failures due to a single-use mock
response being sent multiple times
- added `DynamicResponse` to support constructing mock responses which
can be passed the relevant gRPC request
- changed `MockGuard::wait_until_satisfied` to take `self` by ref rather
than value, since if it consumes self and `wait_until_satisfied` times
out, the `MockGuard` panics in its `Drop` impl, meaning we don't get a
good indication from e.g. a `tokio::time::timeout` as to what went wrong

Most of the new functionality lives in `MockCelestiaAppServer` and
`TestSequencerRelayer`. For the former, all gRPCs except for
`BroadcastTx` and `GetTx` are set up to always return valid-enough
responses - i.e. these don't need to be mounted individually in every
test case. In the case of `MockCelestiaAppServer`, the new functionality
is a combination of support for mounting mock responses and functions
with timeouts to query the `/status`, `/healthz` and `/readyz` http
endpoints of the sequencer-relayer.

## Testing
These changes are tests.

Closes #1008.