refactor(conductor): fetch missing blocks as necessary (#1054)
## Summary
Conductor fetches blocks from the rollup node if they are not in its
cache.

## Background
During initialization and when set to firm-and-soft commitment mode,
Conductor issued a `BatchGetBlocks` RPC to the rollup. When the delta
between firm and soft commitments was too large, this could be a request
for 50k and more elements, putting a lot of pressure on the rollup node
or failing outright because the request exceeded the default 4MB limit
for gRPC requests.

This patch replaces this pre-fetch in favor of an already existing
fallback logic: if conductor finds a Celestia block with the next
desired firm-number, it either looks up the corresponding rollup block
in its cache, or requests it from the rollup. It then updates the
rollup's commitment state.

## Changes
- Remove `BatchGetBlocks` logic from conductor

## Testing
Update the conductor test for fetching missing rollup blocks.