feat(telemetry): register metrics via callback (#1062)
## Summary
Register metrics for a service through a callback on the telemetry
builder.

## Background
`astria_telemetry::Config` should be the central entry point to
configure telemetry and metrics. Before this patch the metrics provider,
buckets, and global labels were set through config, but actual metrics
specific outside of it. With this patch this is done in one step.

## Changes
- add `Config::register_metrics` on telemetry builder
- change all services to pass their initialization function to it. 

## Testing
Must be observed in a dev cluster.