fix(conductor): don't exit on bad Sequencer connection (#1076)
## Summary
Conductor no longer exits if it can't observe the initial height of the
Sequencer network.

## Background
When having soft commitments enabled, Conductor attempted to read the
current height of the Sequencer network during initialization. If this
failed, Conductor would immediately exit rather than retry. In the
Astria dev environment this would happen repeatedly because Conductor's
startup is much faster than Sequencer.

This changes removes this initial height-read entirely because it is not
necessary.

## Changes
- Remove reading the current height of the Sequencer network during
initialization
- Change the latest observed sequencer height in the internal block
stream to be optional: the stream will now only start producing blocks
once a height was externally recorded

## Testing
This change should be transparent to all blackbox tests: they still run
to completion. A unit test was added to check that the mechanism to
return the next height to fetch inside the block-stream returns `None`
if no height was recorded.