feat(sequencer)!: fees go to sudo poa (#1104)
## Summary
Gives block fees to POA authority, and updates some snapshot tests.

## Background
During POA phase of network, we want network fees to go to the authority
account. This allows use to rebate fees to proposers while figuring out
economics.

## Changes
- Fees to sudo instead of proposer.
- Update default genesis for tests. On first pass the snapshot tests
where not failing because the default proposer and the default sudo
address were the same (just a zero address). Updated default genesis so
this should fail in the future. If payee changes.
- Fix test for payee to now be a test for fees going to sudo address.

## Testing
CI. Updated snapshot test which was not working, manually verified that
changing back to proposer as fee recipient will cause test failure.

## Breaking Changelist
- Changes receiver of fees for all blocks.