chore: provide newtypes for verification key and signature (#1111)
## Summary
The `ed25519-consensus` types not already wrapped in newtypes are
wrapped in this PR.

## Background
This is partly for consistency (we recently provided a newtype for the
signing key), partly to avoid having third party types in our core
crate's public API, and partly to support memoizing the address of the
verification key.

## Changes
- Provided newtypes in the `crypto` module of `astria-core` for
`VerificationKey`, `Signature` and `Error`.
- Replaced all usages of the naked `ed25519-consensus` types with the
newtypes.
- Replaced `Address::from_verification_key` with
`VerificationKey::address`.

## Testing
Provided unit tests for the `VerificationKey` to ensure the `PartialEq`,
`Eq`, `PartialOrd`, `Ord` and `Hash` impls are correct given that
they're hand-rolled in order to exclude the memoized address field.

## Related Issues
Closes #1079.