chore(sequencer): refactor mempool (#1092)
## Summary
The mempool was refactored to reduce the visibility, count and scope of
acquired lock guards, and to be implemented in terms of a single
collection rather than two which needed synchronized.

## Background
The mempool has been implemented to be functionally correct, but can
benefit from a refactor to improve safety, performance and complexity
(by avoiding synchronizing two collections).

**NOTE**: This PR is based on top of #1073. Only the final two commits
comprise the new work in this PR.

## Changes
- Having reduced the `BasicMempool` to only holding a single collection,
there seemed little point to retaining it, so the `Mempool` now just
holds the naked priority queue.
- Changed from using a `DoublePriorityQueue` to a `PriorityQueue` since
we don't need the extra functionality, and the latter is slightly more
performant.
- All mempool methods now take `&self` rather than `&mut self`.
- The `RwLock` is no longer locked and unlocked in a tight loop in the
`insert_all` and `remove_all` methods
- `TransactionPriority` has been updated to only hold the nonce diff.
There was a subtle bug in the previous impl since the derived
`PartialEq` and `Eq` didn't correspond to the hand-rolled impls of
`PartialOrd` and `Ord`. This has been fixed and further tests added.
- `SignedTransaction` is internally held in an `Arc` now and passed in
an `Arc` when executing by the app to reduce the number of clones. We
could potentially look to extend this across more of the sequencer code
for the same reason (not in this PR)
- In [the final
commit](https://github.com/astriaorg/astria/commit/fd4402b3b7c1b011b82c3875dd080c7e45d1ba13),
I changed how `update_mempool_after_finalization` was implemented, so
that the mempool no longer has to expose a lock guard. I think it's
generally much safer to not expose the lock or a guard on the lock in a
non-private API. As well as moving the implementation inside mempool, I
added a temporary in-mem cache of current account nonces, since it seems
that there could be many txs all under the same account. As already
noted, this method could be problematic due to holding the mempool lock
for an extended duration, so this might reduce the risk.
 
## Testing
Unit tests extended/added.

## Related Issues
Closes #1083.

---------

Co-authored-by: elizabeth <elizabethjbinks@gmail.com>
Co-authored-by: noot <36753753+noot@users.noreply.github.com>